+++ platforms/python/snips_nlu_rust/nlu_engine.py
@@ -4,15 +4,36 @@ from __future__ import (absolute_import, division, print_function,
 
 import json
 from builtins import object, str
-from ctypes import byref, c_char_p, c_void_p, pointer, string_at, c_char
+from ctypes import byref, c_char_p, c_void_p, pointer, string_at, c_char, c_int
 from pathlib import Path
 
-from snips_nlu_rust.utils import lib, string_pointer
+from snips_nlu_rust.utils import lib, string_pointer, CStringArray
 
 
 class NLUEngine(object):
+    """Python wrapper of the Rust NLU engine
+
+    This wrapper is similar, and can be used in conjunction with the NLU
+    engine defined in the snips-nlu python library
+    (https://github.com/snipsco/snips-nlu), except that it only provides
+    APIs for inference and not for training.
+
+    Examples:
+
+        >>> import io
+        >>> import json
+        >>> from snips_nlu import SnipsNLUEngine as TrainingEngine
+        >>> from snips_nlu_rust import NLUEngine as InferenceEngine
+        >>> with io.open("/path/to/dataset.json") as f:
+        ...     dataset = json.load(f)
+        >>> engine = TrainingEngine().fit(dataset)
+        >>> engine.persist("/path/to/nlu_engine")
+        >>> inference_engine = InferenceEngine(
+        ...     engine_dir="/path/to/nlu_engine")
+        >>> inference_engine.parse("Turn on the lights in the kitchen")
+    """
     def __init__(self, engine_dir=None, engine_bytes=None):
-        exit_code=1
+        exit_code = 1
         self._engine = None
 
         if engine_dir is None and engine_bytes is None:
@@ -25,7 +46,7 @@ class NLUEngine(object):
                               % str(engine_dir))
             self._engine = pointer(c_void_p())
             exit_code = lib.ffi_snips_nlu_engine_create_from_dir(
-                str(engine_dir).encode("utf-8"), byref(self._engine))
+                str(engine_dir).encode("utf8"), byref(self._engine))
         elif engine_bytes is not None:
             self._engine = pointer(c_void_p())
             bytearray_type = c_char * len(engine_bytes)
@@ -37,14 +58,82 @@ class NLUEngine(object):
             raise ImportError('Something wrong happened while creating the '
                               'intent parser. See stderr.')
 
+    def parse(self, query, intents_whitelist=None, intents_blacklist=None):
+        """Extracts intent and slots from an input query
+
+        Args:
+            query (str): input to process
+            intents_whitelist (list of str, optional): if defined, this will restrict the scope of
+                intent parsing to the provided intents
+            intents_blacklist (list of str, optional): if defined, these intents will be excluded
+                from the scope of intent parsing
+
+        Returns:
+            A python dict containing data about intent and slots. See
+            https://snips-nlu.readthedocs.io/en/latest/tutorial.html#parsing for details about the
+            format.
+        """
+        if intents_whitelist is not None:
+            if not all(isinstance(intent, str) for intent in intents_whitelist):
+                raise TypeError(
+                    "Expected 'intents_whitelist' to contain objects of type 'str'")
+            intents = [intent.encode("utf8") for intent in intents_whitelist]
+            arr = CStringArray()
+            arr.size = c_int(len(intents))
+            arr.data = (c_char_p * len(intents))(*intents)
+            intents_whitelist = byref(arr)
+        if intents_blacklist is not None:
+            if not all(isinstance(intent, str) for intent in intents_blacklist):
+                raise TypeError(
+                    "Expected 'intents_blacklist' to contain objects of type 'str'")
+            intents = [intent.encode("utf8") for intent in intents_blacklist]
+            arr = CStringArray()
+            arr.size = c_int(len(intents))
+            arr.data = (c_char_p * len(intents))(*intents)
+            intents_blacklist = byref(arr)
+        with string_pointer(c_char_p()) as ptr:
+            lib.ffi_snips_nlu_engine_run_parse_into_json(
+                self._engine, query.encode("utf8"), intents_whitelist, intents_blacklist,
+                byref(ptr))
+            result = string_at(ptr)
+
+        return json.loads(result.decode("utf8"))
+
+    def get_slots(self, query, intent):
+        """Extracts slots from the input when the intent is known
+
+        Args:
+            query (str): input to process
+            intent (str): intent which the input corresponds to
+
+        Returns:
+            A list of slots. See
+            https://snips-nlu.readthedocs.io/en/latest/tutorial.html#parsing for details about the
+            format.
+        """
+        with string_pointer(c_char_p()) as ptr:
+            lib.ffi_snips_nlu_engine_run_get_slots_into_json(
+                self._engine, query.encode("utf8"), intent.encode("utf8"), byref(ptr))
+            result = string_at(ptr)
+        return json.loads(result.decode("utf8"))
+
+    def get_intents(self, query):
+        """Returns all intents sorted by decreasing confidence scores
+
+        Args:
+            query (str): input to process
+
+        Returns:
+            A list of intents along with their probability. See
+            https://snips-nlu.readthedocs.io/en/latest/tutorial.html#parsing for details about the
+            format.
+        """
+        with string_pointer(c_char_p()) as ptr:
+            lib.ffi_snips_nlu_engine_run_get_intents_into_json(
+                self._engine, query.encode("utf8"), byref(ptr))
+            result = string_at(ptr)
+        return json.loads(result.decode("utf8"))
+
     def __del__(self):
         if self._engine is not None and lib is not None:
             lib.ffi_snips_nlu_engine_destroy_client(self._engine)
-
-    def parse(self, query):
-        with string_pointer(c_char_p()) as ptr:
-            lib.ffi_snips_nlu_engine_run_parse_into_json(
-                self._engine, query.encode("utf-8"), byref(ptr))
-            result = string_at(ptr)
-
-        return json.loads(result.decode("utf-8"))