+++ src/intent_parser/deterministic_intent_parser.rs
@@ -5,8 +5,8 @@ use std::path::Path;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use failure::ResultExt;
-use itertools::Itertools;
+use failure::{format_err, ResultExt};
+use log::{debug, info};
 use regex::{Regex, RegexBuilder};
 use snips_nlu_ontology::{BuiltinEntityKind, IntentClassifierResult, Language};
 use snips_nlu_utils::language::Language as NluUtilsLanguage;
@@ -25,13 +25,14 @@ use crate::utils::{
 };
 
 use super::{internal_parsing_result, IntentParser, InternalParsingResult};
+use itertools::Itertools;
 
 pub struct DeterministicIntentParser {
     language: Language,
     regexes_per_intent: HashMap<IntentName, Vec<Regex>>,
     group_names_to_slot_names: HashMap<String, SlotName>,
     slot_names_to_entities: HashMap<IntentName, HashMap<SlotName, EntityName>>,
-    builtin_scope: Vec<BuiltinEntityKind>,
+    entity_scopes: HashMap<IntentName, (Vec<BuiltinEntityKind>, Vec<EntityName>)>,
     ignore_stop_words: bool,
     shared_resources: Arc<SharedResources>,
 }
@@ -41,6 +42,10 @@ impl DeterministicIntentParser {
         path: P,
         shared_resources: Arc<SharedResources>,
     ) -> Result<Self> {
+        info!(
+            "Loading deterministic intent parser ({:?}) ...",
+            path.as_ref()
+        );
         let parser_model_path = path.as_ref().join("intent_parser.json");
         let model_file = File::open(&parser_model_path).with_context(|_| {
             format!(
@@ -50,7 +55,9 @@ impl DeterministicIntentParser {
         })?;
         let model: DeterministicParserModel = serde_json::from_reader(model_file)
             .with_context(|_| "Cannot deserialize DeterministicIntentParser json data")?;
-        Self::new(model, shared_resources)
+        let parser = Self::new(model, shared_resources);
+        info!("Deterministic intent parser loaded");
+        parser
     }
 }
 
@@ -60,22 +67,35 @@ impl DeterministicIntentParser {
         shared_resources: Arc<SharedResources>,
     ) -> Result<Self> {
         let language = Language::from_str(&model.language_code)?;
-        let builtin_scope = model
+        let entity_scopes = model
             .slot_names_to_entities
             .iter()
-            .flat_map(|(_, mapping)| {
-                mapping
+            .map(|(intent, mapping)| {
+                let builtin_entities = mapping
                     .iter()
                     .flat_map(|(_, entity)| BuiltinEntityKind::from_identifier(entity).ok())
+                    .unique()
+                    .collect();
+                let custom_entities = mapping
+                    .iter()
+                    .flat_map(|(_, entity)| {
+                        if BuiltinEntityKind::from_identifier(entity).is_ok() {
+                            None
+                        } else {
+                            Some(entity.to_string())
+                        }
+                    })
+                    .unique()
+                    .collect();
+                (intent.to_string(), (builtin_entities, custom_entities))
             })
-            .unique()
             .collect();
         Ok(DeterministicIntentParser {
             language,
             regexes_per_intent: compile_regexes_per_intent(model.patterns)?,
             group_names_to_slot_names: model.group_names_to_slot_names,
             slot_names_to_entities: model.slot_names_to_entities,
-            builtin_scope,
+            entity_scopes,
             ignore_stop_words: model.config.ignore_stop_words,
             shared_resources,
         })
@@ -88,17 +108,29 @@ impl IntentParser for DeterministicIntentParser {
         input: &str,
         intents_whitelist: Option<&[&str]>,
     ) -> Result<InternalParsingResult> {
-        Ok(self
+        debug!("Extracting intents and slots with deterministic intent parser...");
+        let result = self
             .parse_top_intents(input, 1, intents_whitelist)?
             .into_iter()
             .next()
+            .and_then(|res| {
+                // return None in case of ambiguity
+                if res.intent.confidence_score < 1.0 {
+                    None
+                } else {
+                    Some(res)
+                }
+            })
             .unwrap_or_else(|| InternalParsingResult {
                 intent: IntentClassifierResult {
                     intent_name: None,
                     confidence_score: 1.0,
                 },
                 slots: vec![],
-            }))
+            });
+        debug!("Intent found: '{:?}'", result.intent.intent_name);
+        debug!("{} slots extracted", result.slots.len());
+        Ok(result)
     }
 
     fn get_intents(&self, input: &str) -> Result<Vec<IntentClassifierResult>> {
@@ -145,61 +177,77 @@ impl DeterministicIntentParser {
         top_n: usize,
         intents: Option<&[&str]>,
     ) -> Result<Vec<InternalParsingResult>> {
-        let opt_intents_set: Option<HashSet<&str>> =
-            intents.map(|intent_list| intent_list.iter().cloned().collect());
-        let builtin_entities = self
-            .shared_resources
-            .builtin_entity_parser
-            .extract_entities(input, Some(self.builtin_scope.as_ref()), true)?
-            .into_iter()
-            .map(|entity| entity.into());
-
-        let custom_entities = self
-            .shared_resources
-            .custom_entity_parser
-            .extract_entities(input, None)?
-            .into_iter()
-            .map(|entity| entity.into());
-
-        let mut matched_entities: Vec<MatchedEntity> = vec![];
-        matched_entities.extend(builtin_entities);
-        matched_entities.extend(custom_entities);
-
-        let (ranges_mapping, formatted_input) =
-            replace_entities(input, matched_entities, get_entity_placeholder);
         let cleaned_input = self.preprocess_text(input);
-        let cleaned_formatted_input = self.preprocess_text(&*formatted_input);
-
         let mut results = vec![];
 
-        for (intent, regexes) in self.regexes_per_intent.iter() {
-            if results.len() == top_n {
-                break;
-            }
-            if !opt_intents_set
-                .as_ref()
-                .map(|intents_set| intents_set.contains(&**intent))
-                .unwrap_or(true)
-            {
-                continue;
-            }
-            for regex in regexes {
-                if let Some(matching_result_formatted) = self.get_matching_result(
-                    input,
-                    &*cleaned_formatted_input,
-                    regex,
-                    intent,
-                    Some(&ranges_mapping),
-                ) {
-                    results.push(matching_result_formatted);
-                } else if let Some(matching_result) =
-                    self.get_matching_result(input, &*cleaned_input, regex, intent, None)
-                {
-                    results.push(matching_result);
-                }
-            }
+        let intents_set: HashSet<&str> = intents
+            .map(|intent_list| intent_list.iter().map(|intent| *intent).collect())
+            .unwrap_or_else(|| {
+                self.slot_names_to_entities
+                    .keys()
+                    .map(|intent| &**intent)
+                    .collect()
+            });
+        let filtered_entity_scopes = self
+            .entity_scopes
+            .iter()
+            .filter(|(intent, _)| intents_set.contains(&***intent));
+
+        for (intent, (builtin_scope, custom_scope)) in filtered_entity_scopes {
+            let builtin_entities = self
+                .shared_resources
+                .builtin_entity_parser
+                .extract_entities(input, Some(builtin_scope.as_ref()), true)?
+                .into_iter()
+                .map(|entity| entity.into());
+
+            let custom_entities = self
+                .shared_resources
+                .custom_entity_parser
+                .extract_entities(input, Some(custom_scope.as_ref()))?
+                .into_iter()
+                .map(|entity| entity.into());
+
+            let mut matched_entities: Vec<MatchedEntity> = vec![];
+            matched_entities.extend(builtin_entities);
+            matched_entities.extend(custom_entities);
+
+            let (ranges_mapping, formatted_input) =
+                replace_entities(input, matched_entities, get_entity_placeholder);
+            let cleaned_formatted_input = self.preprocess_text(&*formatted_input);
+            self.regexes_per_intent
+                .get(intent)
+                .ok_or_else(|| format_err!("No associated regexes for intent '{}'", intent))?
+                .iter()
+                .find_map(|regex| {
+                    self.get_matching_result(
+                        input,
+                        &*cleaned_formatted_input,
+                        regex,
+                        intent,
+                        Some(&ranges_mapping),
+                    )
+                    .or_else(|| {
+                        self.get_matching_result(input, &*cleaned_input, regex, intent, None)
+                    })
+                })
+                .map(|matching_result_formatted| results.push(matching_result_formatted));
         }
-        Ok(results)
+
+        let confidence_score = if results.is_empty() {
+            1.0
+        } else {
+            1.0 / (results.len() as f32)
+        };
+
+        Ok(results
+            .into_iter()
+            .take(top_n)
+            .map(|mut res| {
+                res.intent.confidence_score = confidence_score;
+                res
+            })
+            .collect())
     }
 
     fn preprocess_text(&self, string: &str) -> String {
@@ -363,6 +411,8 @@ mod tests {
     use crate::testutils::*;
 
     use super::*;
+    use crate::entity_parser::builtin_entity_parser::BuiltinEntityParser;
+    use crate::entity_parser::custom_entity_parser::CustomEntityParser;
 
     fn test_configuration() -> DeterministicParserModel {
         DeterministicParserModel {
@@ -384,6 +434,9 @@ mod tests {
                 ],
                 "dummy_intent_4".to_string() => vec![
                     r"^\s*what\s*is\s*(?P<group8>%SNIPSNUMBER%)\s*plus\s*(?P<group8_2>%SNIPSNUMBER%)\s*$".to_string()
+                ],
+                "dummy_intent_5".to_string() => vec![
+                    r"^\s*Send\s*5\s*dollars\s*to\s*john\s*$".to_string(),
                 ],
             ],
             group_names_to_slot_names: hashmap![
@@ -413,6 +466,7 @@ mod tests {
                 "dummy_intent_4".to_string() => hashmap![
                     "dummy_slot_name5".to_string() => "snips/number".to_string(),
                 ],
+                "dummy_intent_5".to_string() => hashmap![],
             ],
             config: DeterministicParserConfig {
                 ignore_stop_words: true,
@@ -490,6 +544,42 @@ mod tests {
     }
 
     #[test]
+    fn test_ambiguous_intent_should_not_be_parsed() {
+        // Given
+        let text = "Send 5 dollars to john";
+        let mocked_builtin_entity_parser = MockedBuiltinEntityParser::from_iter(vec![(
+            text.to_string(),
+            vec![BuiltinEntity {
+                value: "5 dollars".to_string(),
+                range: 5..14,
+                entity: SlotValue::AmountOfMoney(AmountOfMoneyValue {
+                    value: 5.,
+                    precision: Precision::Exact,
+                    unit: Some("dollars".to_string()),
+                }),
+                entity_kind: BuiltinEntityKind::AmountOfMoney,
+            }],
+        )]);
+        let shared_resources = SharedResourcesBuilder::default()
+            .builtin_entity_parser(mocked_builtin_entity_parser)
+            .build();
+        let parser =
+            DeterministicIntentParser::new(test_configuration(), Arc::new(shared_resources))
+                .unwrap();
+
+        // When
+        let intent = parser.parse(text, None).unwrap().intent;
+
+        // Then
+        let expected_intent = IntentClassifierResult {
+            intent_name: None,
+            confidence_score: 1.0,
+        };
+
+        assert_eq!(intent, expected_intent);
+    }
+
+    #[test]
     fn test_parse_intent_with_whitelist() {
         // Given
         let text = "this is a dummy_a query with another dummy_c";
@@ -535,26 +625,22 @@ mod tests {
     #[test]
     fn test_get_intents() {
         // Given
-        let text = "this is a dummy_a query with another dummy_c";
-        let mocked_custom_entity_parser = MockedCustomEntityParser::from_iter(vec![(
+        let text = "Send 5 dollars to john";
+        let mocked_builtin_entity_parser = MockedBuiltinEntityParser::from_iter(vec![(
             text.to_string(),
-            vec![
-                CustomEntity {
-                    value: "dummy_a".to_string(),
-                    resolved_value: "dummy_a".to_string(),
-                    range: 10..17,
-                    entity_identifier: "dummy_entity_1".to_string(),
-                },
-                CustomEntity {
-                    value: "dummy_c".to_string(),
-                    resolved_value: "dummy_c".to_string(),
-                    range: 37..44,
-                    entity_identifier: "dummy_entity_2".to_string(),
-                },
-            ],
+            vec![BuiltinEntity {
+                value: "5 dollars".to_string(),
+                range: 5..14,
+                entity: SlotValue::AmountOfMoney(AmountOfMoneyValue {
+                    value: 5.,
+                    precision: Precision::Exact,
+                    unit: Some("dollars".to_string()),
+                }),
+                entity_kind: BuiltinEntityKind::AmountOfMoney,
+            }],
         )]);
         let shared_resources = SharedResourcesBuilder::default()
-            .custom_entity_parser(mocked_custom_entity_parser)
+            .builtin_entity_parser(mocked_builtin_entity_parser)
             .build();
         let parser =
             DeterministicIntentParser::new(test_configuration(), Arc::new(shared_resources))
@@ -564,15 +650,25 @@ mod tests {
         let intents = parser.get_intents(text).unwrap();
 
         // Then
-        let first_intent = IntentClassifierResult {
-            intent_name: Some("dummy_intent_1".to_string()),
-            confidence_score: 1.0,
-        };
-        assert_eq!(5, intents.len());
-        assert_eq!(&first_intent, &intents[0]);
-        assert_eq!(0.0, intents[1].confidence_score);
-        assert_eq!(0.0, intents[2].confidence_score);
-        assert_eq!(0.0, intents[3].confidence_score);
+        let scores = intents
+            .iter()
+            .map(|res| res.confidence_score)
+            .collect::<Vec<_>>();
+        let expected_scores = vec![0.5, 0.5, 0.0, 0.0, 0.0, 0.0];
+        let intent_names = intents
+            .into_iter()
+            .skip(2)
+            .map(|res| res.intent_name.unwrap_or("null".to_string()).to_string())
+            .sorted()
+            .collect::<Vec<_>>();
+        let expected_intent_names = vec![
+            "dummy_intent_1".to_string(),
+            "dummy_intent_2".to_string(),
+            "dummy_intent_4".to_string(),
+            "null".to_string(),
+        ];
+        assert_eq!(expected_scores, scores);
+        assert_eq!(expected_intent_names, intent_names);
     }
 
     #[test]
@@ -594,6 +690,119 @@ mod tests {
         )]);
         let shared_resources = SharedResourcesBuilder::default()
             .builtin_entity_parser(mocked_builtin_entity_parser)
+            .build();
+        let parser =
+            DeterministicIntentParser::new(test_configuration(), Arc::new(shared_resources))
+                .unwrap();
+
+        // When
+        let intent = parser.parse(text, None).unwrap().intent;
+
+        // Then
+        let expected_intent = IntentClassifierResult {
+            intent_name: Some("dummy_intent_3".to_string()),
+            confidence_score: 1.0,
+        };
+
+        assert_eq!(intent, expected_intent);
+    }
+
+    #[test]
+    fn test_parse_intent_with_entities_from_different_intents() {
+        // Given
+        let text = "Send 10 dollars to John at the wall";
+
+        #[derive(Default)]
+        pub struct MyMockedBuiltinEntityParser;
+
+        impl BuiltinEntityParser for MyMockedBuiltinEntityParser {
+            fn extract_entities(
+                &self,
+                sentence: &str,
+                filter_entity_kinds: Option<&[BuiltinEntityKind]>,
+                _use_cache: bool,
+            ) -> Result<Vec<BuiltinEntity>> {
+                let mocked_builtin_entity_number = BuiltinEntity {
+                    value: "10".to_string(),
+                    range: 5..7,
+                    entity: SlotValue::Number(NumberValue { value: 10. }),
+                    entity_kind: BuiltinEntityKind::Number,
+                };
+                let mocked_builtin_entity_money = BuiltinEntity {
+                    value: "10 dollars".to_string(),
+                    range: 5..15,
+                    entity: SlotValue::AmountOfMoney(AmountOfMoneyValue {
+                        value: 10.,
+                        precision: Precision::Exact,
+                        unit: Some("dollars".to_string()),
+                    }),
+                    entity_kind: BuiltinEntityKind::AmountOfMoney,
+                };
+                if sentence != "Send 10 dollars to John at the wall" {
+                    return Ok(vec![]);
+                }
+                Ok(filter_entity_kinds
+                    .map(|entity_kinds| {
+                        let mut entities = vec![];
+                        if entity_kinds.contains(&mocked_builtin_entity_number.entity_kind) {
+                            entities.push(mocked_builtin_entity_number.clone())
+                        };
+                        if entity_kinds.contains(&mocked_builtin_entity_money.entity_kind) {
+                            entities.push(mocked_builtin_entity_money.clone())
+                        };
+                        entities
+                    })
+                    .unwrap_or_else(|| {
+                        vec![mocked_builtin_entity_number, mocked_builtin_entity_money]
+                    }))
+            }
+        }
+
+        #[derive(Default)]
+        pub struct MyMockedCustomEntityParser;
+
+        impl CustomEntityParser for MyMockedCustomEntityParser {
+            fn extract_entities(
+                &self,
+                sentence: &str,
+                filter_entity_kinds: Option<&[String]>,
+            ) -> Result<Vec<CustomEntity>> {
+                let mocked_custom_entity_1 = CustomEntity {
+                    value: "John".to_string(),
+                    resolved_value: "John".to_string(),
+                    range: 19..23,
+                    entity_identifier: "dummy_entity_1".to_string(),
+                };
+                let mocked_custom_entity_2 = CustomEntity {
+                    value: "the wall".to_string(),
+                    resolved_value: "the wall".to_string(),
+                    range: 27..35,
+                    entity_identifier: "dummy_entity_2".to_string(),
+                };
+                if sentence != "Send 10 dollars to John at the wall" {
+                    return Ok(vec![]);
+                }
+                Ok(filter_entity_kinds
+                    .map(|entity_kinds| {
+                        let mut entities = vec![];
+                        if entity_kinds.contains(&mocked_custom_entity_1.entity_identifier) {
+                            entities.push(mocked_custom_entity_1.clone())
+                        };
+                        if entity_kinds.contains(&mocked_custom_entity_2.entity_identifier) {
+                            entities.push(mocked_custom_entity_2.clone())
+                        };
+                        entities
+                    })
+                    .unwrap_or_else(|| vec![mocked_custom_entity_1, mocked_custom_entity_2]))
+            }
+        }
+
+        let my_mocked_builtin_entity_parser = MyMockedBuiltinEntityParser {};
+        let my_mocked_custom_entity_parser = MyMockedCustomEntityParser {};
+
+        let shared_resources = SharedResourcesBuilder::default()
+            .builtin_entity_parser(my_mocked_builtin_entity_parser)
+            .custom_entity_parser(my_mocked_custom_entity_parser)
             .build();
         let parser =
             DeterministicIntentParser::new(test_configuration(), Arc::new(shared_resources))